refactor: consolidate provenance schemas, helpers, and edit-history endpoints
Move edit-history schemas (FieldChangeSchema, ChangeSetSchema), helpers
(build_edit_history, claims_prefetch, build_sources), and 10 identical
per-entity edit-history endpoints from the catalog app into provenance
where they belong. Replace the duplicated endpoints with a single generic
GET /api/edit-history/{entity_type}/{slug}/ endpoint gated by
link_url_pattern. Also fixes deprecated Django Ninja tuple returns in
provenance/api.py to use Status().
