+++ backend/apps/catalog/api/themes.py
@@ -16,15 +16,13 @@ from .edit_claims import (
     plan_parent_claims,
     validate_scalar_fields,
 )
+from apps.provenance.helpers import build_sources, claims_prefetch
+
 from .helpers import (
-    _build_sources,
-    _build_edit_history,
     _build_rich_text,
-    _claims_prefetch,
     _serialize_title_machine,
 )
 from .schemas import (
-    ChangeSetSchema,
     ClaimSchema,
     HierarchyClaimPatchSchema,
     RichTextSchema,
@@ -66,7 +64,7 @@ def _detail_qs():
         Prefetch("parents", queryset=Theme.objects.active()),
         Prefetch("children", queryset=Theme.objects.active()),
         "aliases",
-        _claims_prefetch(),
+        claims_prefetch(),
         Prefetch(
             "machine_models",
             queryset=MachineModel.objects.active()
@@ -90,7 +88,7 @@ def _serialize_detail(theme) -> dict:
             {"name": t.name, "slug": t.slug} for t in theme.children.order_by("name")
         ],
         "machines": [_serialize_title_machine(pm) for pm in theme.machine_models.all()],
-        "sources": _build_sources(getattr(theme, "active_claims", [])),
+        "sources": build_sources(getattr(theme, "active_claims", [])),
     }
 
 
@@ -175,13 +173,3 @@ def patch_theme_claims(request, slug: str, data: HierarchyClaimPatchSchema):
 
     theme = get_object_or_404(_detail_qs(), slug=theme.slug)
     return _serialize_detail(theme)
-
-
-@themes_router.get("/{slug}/edit-history/", response=list[ChangeSetSchema])
-@decorate_view(cache_control(no_cache=True))
-def get_theme_edit_history(request, slug: str):
-    """Return changeset-grouped edit history with old/new diffs."""
-    from ..models import Theme
-
-    theme = get_object_or_404(Theme, slug=slug)
-    return _build_edit_history(theme)
